feat: store the mp3 files in the cloud and use ffprobe to get metadata
this is just a POC, otherwise terrible idea to run ffprobe during
request/response cycle, it's quite a slow process to loop through the 10
files and run ffprobe on each. Better to have this info in advance so
that we skip the ffprobe part altogether, otherwise, consider caching
the ffproble results so that the next time, we check the cache first
before ffprobing!